Add files via upload
A range slider is a user interface component that enables users to select a value within a specified range by dragging a handle along a track. It provides a visual indication of the current value and allows for easy adjustment by interacting directly with the slider handle.
I found this feature on http://themaltasalarycalculator.com to control the input of salary. It is commonly implemented in web development to provide intuitive controls for adjusting settings, such as volume controls, price ranges in e-commerce websites, or any other numerical input that falls within a defined range.